basic file reorg
